Convert logits dtype to match edge_logits

+++ src/wztarf/models/pretraining/topology_heads.py
@@ -243,7 +243,9 @@ class TopologyHeads(nn.Module):
             edge_logits[
                 b,
                 valid_edges,
-            ] = logits
+            ] = logits.to(
+                dtype=edge_logits.dtype
+            )
 
             predicted_edge_mask[
                 b,
